pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "entropy-reward-experiments"
version = "0.1.0"
description = "Entropy reward experiments for agentic mid-training"
readme = "docs/agentic-mid-training.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "datasets>=2.16.0",
    "peft>=0.7.0",
    "trl>=0.7.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "nltk>=3.8.0",
    "wandb>=0.16.0",
    "tensorboard>=2.15.0",
    "jsonlines>=4.0.0",
    "tqdm>=4.66.0",
    "pandas>=2.1.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0", "ruff>=0.1.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"
